refactor: 검색 결과 페이지 폴더 구조 변경 및 페이지로 네이밍 변경
