+++ apps/client/src/pages/search/page/search-page.tsx
@@ -5,12 +5,12 @@ import { useInfiniteScroll } from '@shared/utils/use-infinite-scroll';
 import PopularSearchSection from '../components/search-home/popular-search-section';
 import RecentFestivalSection from '../components/search-home/recent-festivals-section';
 import RecentSearchSection from '../components/search-home/recent-search-section';
-import SearchResult from '../components/search-result/search-result';
 import {
   useSearchPerformances,
   useSearchRelatedKeyword,
 } from '../hooks/use-search-data';
 import { useSearchLogic } from '../hooks/use-search-logic';
+import SearchResult from './search-result-page';
 
 import * as styles from './search-page.css';
 
@@ -83,7 +83,7 @@ const SearchPage = () => {
         </main>
       )}
 
-      {paramsKeyword && (
+      {!hasArtistResults && paramsKeyword && (
         <SearchResult
           artistData={artistData ?? null}
           performanceCount={performanceCount}
