+++ apps/client/src/pages/search/page/search-result-page.css.ts
